Adding modified files after refactoring
