+++ examples/hand_pose_estimation/processors/standard.py
@@ -1,10 +1,12 @@
 from paz.abstract import Processor
 from ..backend.standard import wrap_dictionary, merge_dictionaries
+from paz.backend.boxes import to_one_hot
 
 
 class WraptoDictionary(Processor):
     """ Wrap the input values to a dictionary with already provided key
     values """
+
     def __init__(self, keys):
         super(WraptoDictionary, self).__init__()
         if not isinstance(keys, list):
@@ -19,8 +21,20 @@ class WraptoDictionary(Processor):
 
 class MergeDictionaries(Processor):
     """ Merge two dictionaries into one"""
+
     def __init__(self):
         super(MergeDictionaries, self).__init__()
 
     def call(self, dicts):
         return merge_dictionaries(dicts)
+
+
+class ToOneHot(Processor):
+    """Extract Hand mask."""
+
+    def __init__(self, num_classes=2):
+        super(ToOneHot, self).__init__()
+        self.num_classes = num_classes
+
+    def call(self, class_indices):
+        return to_one_hot(class_indices, self.num_classes)
